backend/requirements.txt
fastapi
uvicorn
celery
redis
python-multipart
pytest
passlib[bcrypt]
python-jose[cryptography]
boto3
python-dotenv
